JustKnobsConfig: teach install_config_module to handle it.
We had to make a couple design decisions. JustKnobsConfig is now
pickleable, and will record all of it's config when pickled, including
manual overrides, but won't store the evaluated value.

When marshalled portably, JustKnobsConfig will be skipped, unless it's
been explititly set. Same for code.

This will retain the ability to use JustKnobsConfig at runtime, while
respecting explicit setting of the value.

Most of this is test infrastructure. I made a fake config module to test
a number of these config marshallers