config: Modify install_config_module to use a layered approach
This modifies the config system, to use a single mapping of config ->
ConfigEntry and to store the default and user values within them.

We could have used multiple dicts (i.e. user_override and default), but
as we add more fields (justknobs in this PR, perhaps testing and env
variables later), it quickly becomes painful.

There are a couple design decisions we could change.
1) All configs we save store the resolved value - not the default and
   user override seperately
2) All configs we load, apply the resolved value as a user override.

This means that certain complexities of default behvaiour and deletion
(as well as JK), will change if you save + load a config.

ghstack-source-id: bb4f1a6feed336d0d0abec8f2d205d1ecf0c1ef6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/138758