pytorch/feature: Record if parallel compile is enabled
This gets a bit messy, but this appears to be the best spot to make a
true / false decision.

Note that since we're looking at whether or not it's used, if the pool
doesn't warm up within the time it takes for a compile, we will mark the
feature use as false.

ghstack-source-id: c258e8a0b0ec4a40da56bcfbe048446d67fc1331
Pull Request resolved: https://github.com/pytorch/pytorch/pull/141074